Move classes into more suitable locations.
This splits out those files that had two classes in as well.
