+++ source/AndroidManifest.xml
@@ -32,7 +32,7 @@
 
         <!-- Just a bunch of activities here. Needed otherwise android cannot launch them. -->
         <activity
-            android:name="in.joind.Main"
+            android:name="in.joind.activity.Main"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -43,55 +43,55 @@
                 android:resource="@xml/searchable"/>
         </activity>
         <activity
-            android:name=".EventDetail"
+            android:name=".activity.EventDetail"
             android:label="@string/app_name"
             android:launchMode="singleTop">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="in.joind.Main"/>
+                android:value="in.joind.activity.Main"/>
         </activity>
         <activity
-            android:name=".EventTalks"
+            android:name=".activity.EventTalks"
             android:label="@string/app_name"
             android:launchMode="singleTop">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="in.joind.EventDetail"/>
+                android:value="in.joind.activity.EventDetail"/>
         </activity>
         <activity
-            android:name=".EventComments"
+            android:name=".activity.EventComments"
             android:label="@string/app_name"
             android:launchMode="singleTop">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="in.joind.EventDetail"/>
+                android:value="in.joind.activity.EventDetail"/>
         </activity>
         <activity
-            android:name=".EventTracks"
+            android:name=".activity.EventTracks"
             android:label="@string/app_name"
             android:launchMode="singleTop">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="in.joind.EventDetail"/>
+                android:value="in.joind.activity.EventDetail"/>
         </activity>
         <activity
-            android:name=".TalkDetail"
+            android:name=".activity.TalkDetail"
             android:label="@string/app_name"
             android:launchMode="singleTop">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="in.joind.EventTalks"/>
+                android:value="in.joind.activity.EventTalks"/>
         </activity>
         <activity
-            android:name=".TalkComments"
+            android:name=".activity.TalkComments"
             android:label="@string/app_name"
             android:launchMode="singleTop">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="in.joind.TalkDetail"/>
+                android:value="in.joind.activity.TalkDetail"/>
         </activity>
         <activity
-            android:name=".AddComment"
+            android:name=".activity.AddComment"
             android:label="@string/app_name"
             android:launchMode="singleTop">
         </activity>
@@ -101,7 +101,7 @@
             android:launchMode="singleTop">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="in.joind.Main"/>
+                android:value="in.joind.activity.Main"/>
         </activity>
 
         <!-- Authentication service. Used to store details of a user's joind.in account -->
