+++ source/src/in/joind/activity/EventTracks.java
@@ -1,18 +1,13 @@
-package in.joind;
+package in.joind.activity;
 
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.ActionBar;
 import android.util.Log;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.AdapterView.OnItemClickListener;
-import android.widget.ArrayAdapter;
 import android.widget.ListView;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.crashlytics.android.Crashlytics;
@@ -23,6 +18,9 @@ import org.json.JSONObject;
 
 import java.util.ArrayList;
 
+import in.joind.api.DataHelper;
+import in.joind.api.JIRest;
+import in.joind.R;
 import in.joind.adapter.TrackAdapter;
 
 /*
