+++ source/src/in/joind/activity/TalkComments.java
@@ -1,4 +1,4 @@
-package in.joind;
+package in.joind.activity;
 
 /*
  * Displays detailed information about a talk (info, comments etc)
@@ -22,6 +22,9 @@ import org.json.JSONObject;
 
 import java.util.ArrayList;
 
+import in.joind.api.DataHelper;
+import in.joind.api.JIRest;
+import in.joind.R;
 import in.joind.adapter.TalkCommentAdapter;
 
 public class TalkComments extends JIActivity implements OnClickListener {
